Add fixAllCampaignAnalytics functionality and refactor existing analytics code
- Created new package.json for fixAllCampaignAnalytics with necessary dependencies.
- Implemented processCampaignAnalytics function in utils.js to handle campaign analytics processing.
- Removed outdated analytics logic from refreshAllCampaignAnalytics/utils.js to streamline codebase.
