+++ maintenance/fixAllCampaignAnalytics/firebase_connect.js
