+++ maintenance/fixAllCampaignAnalytics/index.js
